Create fragment-tab component
pass activeFragmentId from fragment-tab-list, because this.fragmentsController.activeFragmentId returns 0 in fragment-tab

Early return on _activeFragmentListener when SnippetController.selectedSnippet._id is undefined

Remove click listener from fragment-tab-list, fire it up by click in FragmentTab

Fix to display the code
