+++ src/components.ts
@@ -6,5 +6,6 @@ import "./components/codeEditor";
 import "./components/snippet-list";
 import "./components/search-item";
 import "./components/toast-element";
+import "./components/fragment-tab";
 import "./components/fragment-tab-list";
 import "./components/fragment-title";
